ST-2: Adding alembic to project + small fixes on done
