+++ src/app/app/database/__init__.py
@@ -1,0 +1,4 @@
+import os
+
+# DATABASE_URL = os.environ.get("DATABASE_URL")
+DATABASE_URL = "postgresql://igor:1234@localhost:5432/spotify_tracker_db"
