+++ src/app/app/database/base_class.py
@@ -1,0 +1,13 @@
+from typing import Any
+
+from sqlalchemy.ext.declarative import as_declarative, declared_attr
+
+
+@as_declarative()
+class Base:
+    id: Any
+    __name__: str
+
+    @declared_attr
+    def __tablename__(cls) -> str:
+        return ''.join(["_" + char.lower() if char.isupper() else char for char in cls.__name__])[1:]
