+++ src/app/app/models/user.py
@@ -1,7 +1,7 @@
 from sqlalchemy import String, ForeignKey, Column, Integer
 from sqlalchemy.orm import relationship
 
-from src.app.app.database.base import Base
+from app.database.base import Base
 
 
 class User(Base):
